ecommerce/requirements.txt
alembic==1.11.2
aniso8601==9.0.1
asgiref==3.7.2
blinker==1.6.2
certifi==2023.7.22
click==8.1.6
colorama==0.4.6
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
mysqlclient==2.2.0
packaging==23.1
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
vboxapi==1.0
Werkzeug==2.3.7
Django
Pillow
Djangorestframework
